FIX: Handle wedge case properly
The default implementation in OpenFOAM just mirrors the
vector-properties (i.e. the velocity) at the wedge boundary, which is
wrong in all cases.

The new implementation now evaluates the particle destination (and
Langevin equation) only at the beginning of the time step, rotates the
destination position into the centre plane and the fluctuating particle
velocity is also transformed. This way a particle should *never* hit a
wedge patch.

Signed-off-by: Michael Wild <mwild@ifd.mavt.ethz.ch>

git-svn-id: https://svn.origo.ethz.ch/pdffoam/trunk@98 43a219fd-4765-4d09-99ce-3c4172bfe293