ENH: Added a new gradient-interpolation scheme for position correction
The currently implemented scheme decomposes every cell into tetrahedra
formed by the face centre, two cell vertices and the cell centre. The
tetrahedra act like linear FE elements, with constant gradient.

For the interpolation, the vol field is interpolated first to the faces
and points, and from there the gradient is computed inside the
tetrahedra using the FE formulation.

Signed-off-by: Michael Wild <mwild@ifd.mavt.ethz.ch>

git-svn-id: https://svn.origo.ethz.ch/pdffoam/trunk@187 43a219fd-4765-4d09-99ce-3c4172bfe293