ENH: Refactor the mcLocalTimeStepping classes
Have overloaded correct() functions, one operating on the whole cloud,
the other on a single particle. If not overriden, the former just
iterates over the cloud and calls correct() for each particle, while the
latter does nothing. This way, derived classes can only override the
overload for a single particle.

The virtual function setupInternals() can be overriden to perform
special operations required to be run *once* every time-step. The
correct() overload for a single particle should call this function if
the `prepare' argument is true.

Signed-off-by: Michael Wild <mwild@ifd.mavt.ethz.ch>